Move object difficulty calculation to evaluator methods
